refactor(vite strategy): split getDefaultBasebuildViteConfig and mergeViteConfigs
