+++ src/apps/module/core/strategies/vite/merge.ts
@@ -1,0 +1,31 @@
+import deepAssign from '../../../utils/deepAssign.js'
+import { ConfigEnv, UserConfig } from 'vite'
+import debug from 'debug'
+import { BasebuildCoreConfigs } from '../../types.js'
+import getDefaultBasebuildViteConfig from './vite.config.js'
+
+const log = debug('basebuild:vite:merge')
+
+export const mergeViteConfigs = (configs: BasebuildCoreConfigs) => {
+  return (configEnv: ConfigEnv): UserConfig => {
+    const bbDefaultViteConfig = getDefaultBasebuildViteConfig(configEnv)
+
+    const finalConfig: UserConfig = configs.reduce((mergedConfig, config) => {
+      let buildedConfig = config as UserConfig
+
+      if (typeof config === 'function') {
+        buildedConfig = config?.({
+          ...configEnv,
+          basebuildDefaults: mergedConfig as UserConfig,
+        }) as UserConfig
+      }
+
+      deepAssign(mergedConfig, buildedConfig)
+      return mergedConfig as UserConfig
+    }, bbDefaultViteConfig) as UserConfig
+
+    log('[pre-defineConfig] final vite config', finalConfig)
+    log('[build] rollup options', finalConfig.build?.rollupOptions)
+    return finalConfig
+  }
+}
